Refactor graph executor move logic to execution
