+++ lib/prana/execution/graph_executor.ex
@@ -262,8 +262,7 @@ defmodule Prana.GraphExecutor do
   # Find ready nodes and execute following branch-completion strategy.
   # Uses Execution.__runtime for tracking active paths and executed nodes.
   defp find_and_execute_ready_nodes(execution) do
-    ready_nodes =
-      find_ready_nodes(execution.execution_graph, execution.node_executions, execution.__runtime)
+    ready_nodes = Execution.find_ready_nodes(execution)
 
     if Enum.empty?(ready_nodes) do
       # No ready nodes but workflow not complete - likely an error condition
@@ -356,78 +355,6 @@ defmodule Prana.GraphExecutor do
 
   # This function is no longer needed with depth-based approach
   # defp node_continues_active_branch? - removed
-
-  @doc """
-  Find nodes that are ready to execute based on their dependencies and conditional paths.
-
-  A node is ready if:
-  1. It hasn't been executed yet (not in completed node executions)
-  2. All its input dependencies have been satisfied
-  3. It's reachable from completed nodes or is an entry node
-  4. It's on an active conditional execution path (for conditional branching)
-
-  ## Parameters
-
-  - `execution_graph` - The ExecutionGraph containing nodes and dependencies
-  - `completed_node_executions` - List of completed NodeExecution structs
-  - `execution_context` - Current execution context with conditional path tracking
-
-  ## Returns
-
-  List of Node structs that are ready for execution.
-  """
-  @spec find_ready_nodes(ExecutionGraph.t(), map(), map()) :: [Node.t()]
-  def find_ready_nodes(%ExecutionGraph{} = execution_graph, node_executions, execution_context) do
-    # Get active nodes from execution context
-    active_nodes = execution_context["active_nodes"] || MapSet.new()
-
-    # Extract completed node IDs from map structure for dependency checking
-    completed_node_ids =
-      node_executions
-      |> Enum.map(fn {node_key, executions} -> {node_key, List.last(executions)} end)
-      |> Enum.filter(fn {_, exec} -> exec.status == :completed end)
-      |> MapSet.new(fn {node_key, _} -> node_key end)
-
-    # Only check active nodes instead of all nodes
-    active_nodes
-    |> Enum.map(fn node_key -> execution_graph.node_map[node_key] end)
-    |> Enum.reject(&is_nil/1)
-    |> Enum.filter(fn node ->
-      dependencies_satisfied?(node, execution_graph, completed_node_ids)
-    end)
-  end
-
-  # Check if all input ports for a node are satisfied (port-based logic)
-  defp dependencies_satisfied?(node, execution_graph, completed_node_ids) do
-    # Get input ports for this node
-    input_ports =
-      case Execution.get_action_input_ports(node) do
-        {:ok, ports} -> ports
-        # fallback to default
-        _error -> ["input"]
-      end
-
-    # For each input port, check if at least one source connection is satisfied
-    Enum.all?(input_ports, fn input_port ->
-      input_port_satisfied?(node.key, input_port, execution_graph, completed_node_ids)
-    end)
-  end
-
-  # Check if a specific input port is satisfied (at least one source available)
-  defp input_port_satisfied?(node_key, input_port, execution_graph, completed_node_ids) do
-    # Get all incoming connections for this node and port
-    incoming_connections = Execution.get_incoming_connections_for_node_port(execution_graph, node_key, input_port)
-
-    # If no incoming connections, port is satisfied (no dependencies)
-    if Enum.empty?(incoming_connections) do
-      true
-    else
-      # At least one source node must be completed
-      Enum.any?(incoming_connections, fn conn ->
-        MapSet.member?(completed_node_ids, conn.from)
-      end)
-    end
-  end
 
   # Execute a single node with middleware events using unified execution architecture
   defp execute_single_node_with_events(node, execution) do
